+++ src/Determinan_EK.java
@@ -1,13 +1,13 @@
 public class Determinan_EK {
-    public static float determinant(float[][] matrix) {
+    public static double determinant(double[][] matrix) {
         int n = matrix.length;
-        float det = 0;
+        double det = 0;
         if (n == 1) {
             det = matrix[0][0];
         } else if (n == 2) {
             det = matrix[0][0] * matrix[1][1] - matrix[0][1] * matrix[1][0];
         } else {
-            float[][] submatrix = new float[n - 1][n - 1];
+            double[][] submatrix = new double[n - 1][n - 1];
             for (int j = 0; j < n; j++) {
                 for (int i = 1; i < n; i++) {
                     for (int k = 0; k < n; k++) {
